+++ test/Main.hs
@@ -4,6 +4,7 @@ import AcyclicitySpec qualified
 import AlignmentInvariantSpec qualified
 import BindingSharedAbstractionSpec qualified
 import BindingSpec qualified
+import BackendIRSpec qualified
 import CanonicalizerSpec qualified
 import Constraint.SolvedSpec qualified
 import ConstraintGenSpec qualified
@@ -73,6 +74,7 @@ main = do
         die "PresolutionSpec was not wired into the test harness."
     SolveSpec.spec
     ScopeSpec.spec
+    BackendIRSpec.spec
     ProgramSpec.spec
     ResolvedSymbolSpec.spec
     PipelineSpec.spec